Move histogram code to histogram.go and add histogram_test.go file
